[chore] Move es/spanstore/dbmodel to internal directory
