+++ plugin/storage/es/spanstore/internal/dbmodel/from_domain.go
