+++ plugin/storage/es/spanstore/internal/dbmodel/from_domain_test.go
