+++ plugin/storage/es/spanstore/internal/dbmodel/to_domain.go
