+++ plugin/storage/es/spanstore/service_operation.go
@@ -17,7 +17,7 @@ import (
 
 	"github.com/jaegertracing/jaeger/pkg/cache"
 	"github.com/jaegertracing/jaeger/pkg/es"
-	"github.com/jaegertracing/jaeger/plugin/storage/es/spanstore/dbmodel"
+	"github.com/jaegertracing/jaeger/plugin/storage/es/spanstore/internal/dbmodel"
 )
 
 const (
